refactor: split agent k8s resource modules
